Move tests outside of the package to own top level directory
This avoids installing tests to end users production systems. These are
often unwanted in the site-packages directory and unused by end users.
The tests are still included in the source distribution (through
MANIFEST.in) and can be run in exactly the same way as before.

This follows the "Tests outside application code" scheme described in
the pytest documentation

https://docs.pytest.org/en/latest/goodpractices.html#choosing-a-test-layout-import-rules