Add tox.ini to allowing running the full test matrix locally
To run the full test matrix, a contributor can now simply use the
command:

    $ tox

This will run all supported Python versions including PyPy. It finishes
with:

    py35: commands succeeded
    py36: commands succeeded
    py37: commands succeeded
    py38: commands succeeded
    pypy3: commands succeeded
    congratulations :)

Travis now uses tox as the test entrypoint so tests continue run the
same locally and on Travis.

This replaces the deprecated command 'python setup.py test'. This was
deprecated in setuptools v41.5.0 (27 Oct 2019) and will be removed in a
future version. The tests_require option is also deprecated.

As setuptools_required was only being used to run tests, it has also
been cleaned up.