Improve testing speed
We don’t require coverage each time we launch tests. We keep it in the list of
dependencies (as it’s always useful to get coverage when needed), but we don’t
use the option by default. It doesn’t change the hard requirements for having
passing tests, because coverage wasn’t automatically checked.

We also use xdist to launch parallel tests on multicore CPUs. Let’s save some
time and some energy (the world is burning 🔥) for our CI.